Port support for cons in internal_equal. (#1118)
However, there was a snag. It turns out `TailsIter` is over eager when
it comes to checking the end condition. In C there is a function called `CHECK_LIST_END`
used after `FOR_EACH_TAIL` for example `CHECK_LIST_END (sequence, sequence);`. It ensures
that the list was fully walked or at least exited early due to a decision and not because the last item checked was not a CONS cell. If the last item was not a proper cons cell like say `(0 . 0)` then `CHECK_LIST_END` will complain. But we do not always want that.

This is the beginning of a refactor/reorganization of the LispCons iterators. `TailsIter::new` now takes an extra parameter which contains the error to raise when the check fails or None. For this PR the code has been left as it was and the lack of check must be requested with `iter_tails_noendchecked`. In the near future this function will replace `iter_tails` and a new `iter_tails_endchecked` will be added. However, this change requires going through all of our ported code and finding the places that need to be updated. To prepare for that there is a `iter_tails_v2()` which exposes two enums to controls whether end checking is done and whether circular list checking is done.

As a further step towards clarity and simplification the extra `unchecked` iterators have been removed. Their functionality is now available using the enum controls.

To reiterate -- the existing behavior has not been changed. If the code calls `iter_tails()` then the behavior before this PR continues. However, this PR provides all of the tools needed to update the tree.