Fix base64 decoding into multibyte string
When we're told decode into a multibyte string, we don't want to
restore the raw bytes and neither do we want to stuff the bytes into a
utf-8 string. Emacs comes with its own encoding which was lost when
converting the code to Rust.

Perform this encoding step when dealing with multibyte output and set
the tests against outputs taken from running the upstream
implementation.