Skip running bindgen if C source hasn't changed
The cargo build script performs a few different tasks so we run it
whenever any of the source code has changed.

A particularly expensive task is running bindgen to generate Rust code
from the C code automatically. This can take minutes and we run it
even if only Rust code has changed, which would not produce a change.

Teach the Makefile to keep track of a hash for the C source code and
tell the cargo build script. Teach the build script to skip running
bindgen unless the Makefile has determined that it needs to be run.