Avoid that unwind_format_mode_line messes up buffer points (Bug#32777)
* src/xdisp.c (format_mode_line_unwind_data): Before
temporarily selecting a window on another frame, separately save
the point of that window's buffer too.
(unwind_format_mode_line): After undoing the temporary
selection of a window on another frame, separately restore the
buffer point of that window.  This is needed since the
operation that deselects that window will have stored back the
point of that window into its buffer's point which is wrong
since that window was never "officially" selected (Bug#32777).