If the Rust name is not the expected C name we have to set it.
The function name exported to C is the Rust name with a capital F
prepended. If this is not the intended name then we have to set the
correct name with `c_name`. The value of `name` is _NOT_ used because
it could be an invalid C (and Rust) name like for instance `+`.