If the Rust name is not the expected C name we have to set it. (#1205)
The function name exported to C is the Rust name with a capital F
prepended. If this is not the intended name then we have to set the
correct name with c_name. The value of name is NOT used because
it could be an invalid C (and Rust) name like for instance +.

Also, this PR begins standardizing our naming. Rust versions of a DEFUN
which cannot use the orignal name are named foo_lisp.