Rewrite as_type methods in terms of From (#1221)
Currently a lot of type conversions are written as methods on
LispObject, and the From trait implementations are written in terms of
those. It should be the other way around, so move the as_type logic
into the From implementations and make the as_type functions just do
casts. This will make it easier to remove them entirely in the future,
allowing the type system to do more work.

A few as_type functions, like as_window_configuration_or_error, were
only used in a few places, so those were cut entirely.