Port yes-or-no-p (#1232)
The first commit changes some `LispObjects` to `bool` which I noticed as I was porting many `Qnil`s. Then we port `yes-or-no-p`. The C version checks the length of the input but it just seems like a microoptimisation we probably don't care about.

We compare the strings as byte slices and it seems like it'd be a good follow-up to allow comparison of `LispStringRef` and strings or byte slice via `PartialEq`.

Annoyingly, the parameter is allowed to be `nil` upstream which is treated as an empty string but we don't do that here.