Fix error on describe-function for adviced builtin function
`subr-lang` gets `wrong-type-argument` for adviced builtin function
because `subr-lang` expects an argument is a subroutine but adviced
function is no longer regarded as subroutine.

Without this change, `describe-function` signals an error if the
argument is an adviced subroutine such as `rename-buffer`.
Fortunately, in help-fns.el there is an analyzer of such functions
which enables us to get real definition and to call `subr-lang`.