* doc/misc/eshell.texi (Built-ins): Fix alias description
Dear eamcs developers, eshells current documentation first states
that alias definitions are not saved to an alias file, later that
they are saved to an alias file.  I tested it and the latter is
correct.

Please find attached a patch which fixes this.

Thanks for working on emacs which is really great, Gregor

>From 1fe51cc769ab7a30d0896fb3d6105c0561243fa7 Mon Sep 17 00:00:00 2001
From: Gregor Zattler <telegraph@gmx.net>
Date: Wed, 13 Feb 2019 20:19:38 +0100
Subject: * doc/misc/eshell.texi (Built-ins): Fix alias description.

Bring description of built-in 'alias' in line with (info "(eshell) Aliases"),
which describes the actual behaviur.