Correct test
The test `xref-elisp-test-find-defs-defun-c` tests if it correctly
determines that a known C builtin is in fact a C builtin. The builtin
it was using to do this was `get-buffer-create`, but since that's now
a rust builtin instead of a C one, the test failed. We should probably
add a xref-elisp-test-find-defs-defun-rust` test.