Move temp_charset_work to its own header.
This static struct is only used internally in srt/charset.c,
and we need it to be available in a header if we want to port
charset.c  functions to Rust.

Adding it to an existing header like lisp.h or charset.h causes
warnings for all other files other than charset.c that include the
header, so I've moved it to its own header and included that in
charset.c

This temporary header can be safely removed from the codebase
once all charset.c functions have been ported.