Fix temp_charset_work declaration
Due to me not quite understanding how static works in C, the struct
was invisible to anything outside the header file it was declared in.

Now it's technically public, but since only charset.c includes the header
then it's still invisible to the rest of Emacs' C code and accessible
to Rust.

When all other charset.c functions that need this struct have been
ported, we can move its declaration to charset.rs and delete the
temporary header