Revert "Remove font.c code commented out for a decade"
This reverts commit 64d0cd9810af6bd0c378fc6bc666c76ddfa97e40.

Rationale: any font-related code and comments, even if unused
for decades, serves as important source of useful information
in an area of Emacs code that is notoriously under-documented.

Please do NOT remove this stuff until we have an active
expert in this are on board, who will then decide whether
this can be retired.