Make get_param more idiomatic
If we assume the get_param method will never be called directly by C, we
can make it a little more idiomatic, and let the wrapper function take
care of returning the type C expects.

My other assumption here is that we want anything that could return
Qnil or something else in C to return an Option in Rust. Please correct
me if that's wrong./