Add minimal interval length logic.
This code stands in for the TOTAL_LENGTH and LENGTH macros defined in
intervals.h. The most correct approach here would be to implement a
LispIntervalRef type that properly encapsulated this logic as methods,
but this is a bit outside the scope of this specific branch. I'll do it
separately.